+++ packages/carbon-graphs/tests/unit/controls/Scatter/ScatterPanning-spec.js
@@ -393,5 +393,26 @@ describe('Scatter - Panning', () => {
         expect(legendItem.getAttribute('aria-current')).toBe('true');
       });
     });
+    describe('when the same data is passed on multiple clicks of panning', () => {
+      it('Range modified should be false', () => {
+        const panData = {
+          key: 'uid_1',
+          values: [
+            {
+              x: '2016-03-03T12:00:00Z',
+              y: 0,
+            },
+            {
+              x: '2016-04-03T12:00:00Z',
+              y: 20,
+            },
+          ],
+        };
+        graphDefault.reflow(panData);
+        graphDefault.reflow(panData);
+        graphDefault.reflow(panData);
+        expect(graphDefault.config.axis.y.dataRange.isRangeModified).toEqual(false);
+      });
+    });
   });
 });